Changed: Use ShortcutManager APIs to create pinned shortcut on android >=8
This is better way to create shortcuts since launcher does not get access to shortcut data of the app and android itself stores them. It may even solve problems on some launchers, mainly Samsung launcher where randomly all shortcuts get renamed to the app name, considering it will be queuering android for the shortcut names at startup.

The only issue with this is that shortcuts may have an app badge on the icon as well, depending on launcher. This is likely a safety mechanism to know which app created the shortcut in case a malicious app creates shortcuts with icons of a different app. This is not controlled by termux-widget but the launcher itself, so contact launcher devs for an option to remove them.

For android <8, static shortcuts will still be used since pinned shortcuts are not supported.

For more info on shortcut types, check https://github.com/agnostic-apollo/TaskerLauncherShortcut#shortcut-types

Related issue #28