Release: v0.13.1
The `versionCode` has been bumped to `1000` so that users who have installed from F-Droid or GitHub should not have the app attempted to be updated by Google PlayStore and failing and also shown in PlayStore app updates list due to non-collaborative `v0.14` app release on PlayStore that set the `versionCode` higher than the latest F-Droid or GitHub `13` release. Unlike F-Droid, PlayStore does not check for difference in app APK signature before attempting to download and then failing to install due to signature mismatch.

- https://github.com/termux/termux-app/discussions/4000
- https://github.com/termux/termux-app/issues/4012
- https://github.com/termux-play-store/termux-apps/commit/e51f8d2d636c4f016b9fecf60cbe0957c2750377#diff-afedc8b49d567d2de57705052be9f6521366702c15f6ea32b2ae45174cef8221