Fixed: Show `Open Termux:Widget app to refresh widget` message instead of `No files in ~/.shortcuts/` message when initial layout of widget is being shown so that it is not confused with shortcuts directory being empty
In this stuck state which may happen after app updates or if app process gets killed without being restarted by Android, the widgets must be refreshed by starting the Termux:Widget app or with the refresh button in the main activity.

If the widget was successfully refreshed/updated and shortcuts directory was empty, only then the `No files in ~/.shortcuts/` message will be shown.